+++ Resources/Private/JavaScript/Host/Selectors/CR/Images/index.js
@@ -1,22 +1,14 @@
-import {$get, $set} from 'plow-js';
-import {createSelector, defaultMemoize} from 'reselect';
+import {$get} from 'plow-js';
+import {createSelector} from 'reselect';
 
-import {
-    byNameSelector as nodeTypeByNameSelector,
-    subTypesSelector
-} from '../NodeTypes/index';
-
-// TODO: create selector here?
 export const imageByUuid = uuid => $get(['cr', 'images', 'byUuid', uuid]);
 
-
-// TODO: selector is somehow broken?!
+// TODO: selector is somehow broken; e.g. not called?
 export const imageLookup = createSelector(
     [
         $get(['cr', 'images', 'byUuid'])
     ],
     images => imageUuid => {
-        console.log("Image Lookup: ", imageUuid, images, $get(imageUuid, images));
-        return $get(imageUuid, images)
+        return $get(imageUuid, images);
     }
 );
